change WINHEIGHT, WINWIDTH, code optimized

+++ src/collision_check.c
@@ -6,13 +6,13 @@
 /*   By: minsepar <minsepar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/08 22:51:37 by minsepar          #+#    #+#             */
-/*   Updated: 2024/04/09 13:12:38 by minsepar         ###   ########.fr       */
+/*   Updated: 2024/04/10 13:21:20 by minsepar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-void	dir_y_check_p(t_map *map, t_user *user, double new_displacement_y)
+void	dir_y_check_p(t_map *map, t_user *user, float new_displacement_y)
 {
 	// printf("yp\n");
 	// printf("y: [%f] x: [%f]\n", user->y + new_displacement_y + HITBOX, user->x);
@@ -23,7 +23,7 @@ void	dir_y_check_p(t_map *map, t_user *user, double new_displacement_y)
 		user->y += new_displacement_y;
 }
 
-void	dir_y_check_n(t_map *map, t_user *user, double new_displacement_y)
+void	dir_y_check_n(t_map *map, t_user *user, float new_displacement_y)
 {
 	// printf("yn\n");
 	// printf("y: [%f] x: [%f]\n", user->y + new_displacement_y + HITBOX, user->x);
@@ -34,7 +34,7 @@ void	dir_y_check_n(t_map *map, t_user *user, double new_displacement_y)
 		user->y += new_displacement_y;
 }
 
-void	dir_x_check_p(t_map *map, t_user *user, double new_displacement_x)
+void	dir_x_check_p(t_map *map, t_user *user, float new_displacement_x)
 {
 	// printf("xp\n");
 	// printf("y: [%f] x: [%f]\n", user->y, user->x + new_displacement_x - HITBOX);
@@ -45,7 +45,7 @@ void	dir_x_check_p(t_map *map, t_user *user, double new_displacement_x)
 		user->x += new_displacement_x;
 }
 
-void	dir_x_check_n(t_map *map, t_user *user, double new_displacement_x)
+void	dir_x_check_n(t_map *map, t_user *user, float new_displacement_x)
 {
 	// printf("xn\n");
 	// printf("y: [%f] x: [%f]\n", user->y, user->x + new_displacement_x - HITBOX);
